requirements.txt
Django==3.2.16
djangorestframework==3.14.0

os==??

psycopg2==2.8.6

asincio==??
aiogram==3.26.0